rename the lib folder to modules
